@@ -1,6 +1,6 @@
 import React from "react";
-import { makeStyles } from '@material-ui/core/styles';
-import Button from '@material-ui/core/Button';
+import { makeStyles } from '@mui/styles';
+import Button from '@mui/material/Button';
 import { ErrorsBlock, Field, Input } from "reacto-form-inputs";
 import useReactoForm from "reacto-form/esm/useReactoForm";
 import validator from "./formValidator";
